requirements.txt
async-generator==1.10
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
cryptography==36.0.2
flake8==4.0.1
h11==0.13.0
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.9
wsproto==1.1.0
